new_run_id {pg}	R Documentation

Generate new run id

Description

Query data base for a new run id from sequence, storing also timestamp of the run id obtained.

Usage

new_run_id(conn = getOption("pg.conn"), .log = getOption("pg.log", TRUE))

Arguments

conn
active connection to postgres database, by default 'getOption("pg.conn")'.

.log
logical default 'getOption("pg.log",TRUE)' decides if call is logged using *logR*.

[Package pg version 0.1.9 Index]